requirements.txt

# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# ML Frameworks
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1

# RL Libraries
gym==0.26.2
stable-baselines3==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0

# Utilities
requests==2.31.0
python-json-logger==2.0.7
